Generify SpringChoreographer
This begins to let Rebound be non-dependent on Android. I'd like to use
this in Processing, but the frame callback system is different there.

This commits creates a SpringChoreographer interface and creates two
Android implementations of this interface.

Follow-up commits could then split out the Android-specific code in its
own BUCK module, and then also introduce other modules for other
platforms, like Processing.